# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Artifact Registry Service
  version: "0.1.0"
  description: |
    General-purpose artifact distribution for internal services and edge devices.
    Artifacts are immutable once uploaded. Channels (stable, canary, …) are
    mutable pointers to a specific version and support single-step rollback.

    ## Webhook events

    State-changing operations emit webhook events to all configured `WEBHOOK_URLS`.

    ### Event envelope

    ```json
    {
      "event_id":      "<uuidv4>",
      "event_type":    "artifact.uploaded | channel.promoted | channel.rollback",
      "timestamp":     "<RFC3339 UTC>",
      "aggregate_key": "<artifact_id or channel name>",
      "data":          { ... },
      "signature":     "<hex HMAC-SHA256>"
    }
    ```

    ### Signature verification

    The signature covers the envelope with `"signature":""`. To verify:
    1. Parse the received JSON.
    2. Set `signature` to `""`.
    3. Re-marshal (field order is deterministic — struct order).
    4. Compute `HMAC-SHA256(secret, remarshaled_bytes)`.
    5. Compare hex result to the `X-Signature` request header.

    ### Ordering guarantee: **NONE**

    Webhook delivery is **asynchronous**. Delivery order is **not guaranteed**.
    Two events for the same entity may arrive out of order at the consumer.

    Consumers **MUST** treat webhook events as notifications, not as
    authoritative ordered state transitions. When ordering matters, consumers
...

...

...

...

...

...

...
servers:
...

...
security:
  - BearerAuth: []   # applied globally; overridden to [] on /health
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API key passed as `Authorization: Bearer <key>`.

        Three additive scopes, configured via env vars:
        - **read**  — GET artifact and channel endpoints (ARS_READ_KEYS)
        - **write** — upload artifacts; implies read (ARS_WRITE_KEYS)
        - **admin** — promote and rollback; implies write + read (ARS_ADMIN_KEYS)

        Set `ARS_AUTH_DISABLED=true` to bypass enforcement in dev mode.
  schemas:
...
    ChannelPointer:
      type: object
      required:
        - channel
        - type
        - name
        - version
        - artifact_id
        - promoted_at
      properties:
        channel:
          type: string
          example: stable
        type:
          type: string
          example: model
        name:
          type: string
          example: fraud-detector
        version:
          type: string
          pattern: "^v[0-9]+\\.[0-9]+\\.[0-9]+$"
          example: "v1.2.3"
        artifact_id:
          type: string
          format: uuid
        previous_version:
          type: string
          pattern: "^v[0-9]+\\.[0-9]+\\.[0-9]+$"
          description: Absent when no prior promotion exists.
          example: "v1.2.2"
        promoted_at:
          type: string
          format: date-time
        promoted_by:
          type: string
          description: Value of the X-Promoted-By request header, if provided.
          example: "ci-pipeline"
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string

  # ── Reusable responses ─────────────────────────────────────────────────────
  responses:
    Unauthenticated:
      description: Missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "missing or invalid API key"
    Forbidden:
      description: >
        Valid key but insufficient scope. The required scope is named in the
        error message.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            read_required:
              value:
                error: "insufficient scope — required: read"
            write_required:
              value:
                error: "insufficient scope — required: write"
            admin_required:
              value:
                error: "insufficient scope — required: admin"
...
  parameters:
    ArtifactType:
      name: type
      in: path
      required: true
      schema:
        type: string
      description: Artifact type (e.g. model, config, binary, rules).
      example: model
    ArtifactName:
      name: name
      in: path
      required: true
      schema:
        type: string
      example: fraud-detector
...
    ChannelName:
      name: channel
      in: path
      required: true
      schema:
        type: string
      description: Distribution channel name (e.g. stable, canary).
      example: stable
...
paths:
...
  /channels/{channel}/state/{type}/{name}:
    get:
      operationId: getChannelState
      summary: Get current channel state for an artifact
      tags: [channels]
      x-required-scope: read
      parameters:
        - $ref: "#/components/parameters/ChannelName"
        - $ref: "#/components/parameters/ArtifactType"
        - $ref: "#/components/parameters/ArtifactName"
      responses:
        "200":
          description: Current channel pointer.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelPointer"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: No version has been promoted to this channel.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /channels/{channel}/history/{type}/{name}:
    get:
      operationId: getChannelHistory
      summary: Get full promotion history for a channel/artifact pair
      description: >
        Returns all channel pointer records in reverse chronological order
        (most recent first). Returns an empty array when no promotions exist.
      tags: [channels]
      x-required-scope: read
      parameters:
        - $ref: "#/components/parameters/ChannelName"
        - $ref: "#/components/parameters/ArtifactType"
        - $ref: "#/components/parameters/ArtifactName"
      responses:
        "200":
          description: Channel history (may be empty).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChannelPointer"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ── Admin endpoints ───────────────────────────────────────────────────────
...